hyprland: Handle external screens
Using SUPER+arrow makes Hyprland to change focus between monitors.
By default, a workspace is attached to the monitor that has the focus.
Then, when a monitor is plugged, nothing shows up on this screen until the focus is made and a new workspace is created.

In addition, the Tuxedo doesn't seem to handle external screens at a refresh rate greater than 60Hz even at 1080p, which is a bit of a shame...

Signed-off-by: LalieA <83229713+LalieA@users.noreply.github.com>